requirements.txt
pandas
numpy
tqdm
fastparquet
openpyxl
xlrd
kaleido
python-dotenv
requests
ipykernel
statsmodels